Refactor Token class to be DTO
